requirements.txt
Flask==2.3.3
pytrends==4.9.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
flask-wtf==1.2.1
wtforms==3.1.0
python-dotenv==1.0.0
facebook-sdk==3.1.0
schedule==1.2.0